+++ CarsAdviser/Forms/PreferencesForm.cs
@@ -20,6 +20,7 @@ namespace CarsAdviser.Forms
         public List<Cars> similarToPreferences;
         private List<Cars> selectedPreferences = new List<Cars>();
         private static Logger logger = LogManager.GetCurrentClassLogger();
+        private Helper helper = new Helper();
         public PreferencesForm(AccountForm parentForm, int userId)
         {
             InitializeComponent();
@@ -33,6 +34,7 @@ namespace CarsAdviser.Forms
         {
             try
             {
+                prefPanel.Controls.Clear();
                 logger.Info("Загрузка случайных автомобилей");
                 var randomCars = new List<Cars>();
 
@@ -43,23 +45,23 @@ namespace CarsAdviser.Forms
                                                                .ThenInclude(cm => cm.Cars_Stamp)
                                                                .Where(up => up.Users_id == userId)
                                                                .Select(up => up.Cars)
-                                                               .Take(4)
+                                                               .Take(10)
                                                                .ToList();
 
                     randomCars.AddRange(userPreferencesCars);
 
-                    if (randomCars.Count < 4)
+                    if (randomCars.Count < 10)
                     {
 
                         var allCars = context.Cars.Include(c => c.Cars_Model)
                                                   .Include(c => c.Cars_Stamp)
                                                   .ToList();
 
-                        randomCars = allCars.OrderBy(x => Guid.NewGuid()).Take(4).ToList();
+                        randomCars = allCars.OrderBy(x => Guid.NewGuid()).Take(10).ToList();
 
-                        var remainingCount = 4 - randomCars.Count;
+                        var remainingCount = 10 - randomCars.Count;
 
-                        var remainingCars = randomCars.Skip(4).ToList();
+                        var remainingCars = randomCars.Skip(10).ToList();
 
                         randomCars.AddRange(remainingCars);
                     }
@@ -69,65 +71,15 @@ namespace CarsAdviser.Forms
                         var allCars = context.Cars.Include(c => c.Cars_Model)
                                                    .Include(c => c.Cars_Stamp)
                                                    .OrderBy(x => Guid.NewGuid())
-                                                   .Take(4)
+                                                   .Take(10)
                                                    .ToList();
 
                         randomCars.AddRange(allCars);
                     }
 
-                    for (int i = 0; i < randomCars.Count; i++)
+                    foreach (var car in randomCars)
                     {
-                        var car = randomCars[i];
-                        Guna2Panel carPanel = this.Controls.Find($"carPanel{i + 1}", true).FirstOrDefault() as Guna2Panel;
-                        if (carPanel != null)
-                        {
-                            carPanel.Visible = true;
-
-                            Guna2PictureBox carPicture = carPanel.Controls.Find($"carPictureBox{i + 1}", true).FirstOrDefault() as Guna2PictureBox;
-                            if (carPicture != null)
-                            {
-                                if (car.Photo_1 != null)
-                                {
-                                    carPicture.Image = Image.FromFile(car.Photo_1);
-                                }
-                                else
-                                {
-                                    carPicture.Image = Properties.Resources.noAuto;
-                                }
-                            }
-
-                            Guna2PictureBox carBrandPicture = carPanel.Controls.Find($"carBrandPictureBox{i + 1}", true).FirstOrDefault() as Guna2PictureBox;
-                            if (carBrandPicture != null)
-                            {
-                                carBrandPicture.Image = Image.FromFile(GetStampImageLocation(car.Cars_Stamp.Stamp));
-                            }
-
-                            Label carName = carPanel.Controls.Find($"carNameLabel{i + 1}", true).FirstOrDefault() as Label;
-                            if (carName != null)
-                            {
-                                carName.Text = $"{car.Cars_Stamp.Stamp} {car.Cars_Model.Model}";
-                            }
-
-                            Label carYear = carPanel.Controls.Find($"carYearLabel{i + 1}", true).FirstOrDefault() as Label;
-                            if (carYear != null)
-                            {
-                                carYear.Text = $"{car.Year}";
-                            }
-
-                            Guna2CheckBox check = carPanel.Controls.Find($"preferCheckBox{i + 1}", true).FirstOrDefault() as Guna2CheckBox;
-                            if (check != null)
-                            {
-                                check.Tag = car.ID;
-                                if (userPreferencesCars.Any(upc => upc.ID == car.ID))
-                                {
-                                    check.Checked = true;
-                                }
-                                else
-                                {
-                                    check.Checked = false;
-                                }
-                            }
-                        }
+                        LoadCurrentCarPreference(car);
                     }
                 }
             }
@@ -137,45 +89,69 @@ namespace CarsAdviser.Forms
                 MessageBox.Show($"{Local.databaseConnectionError}: {ex.Message}", Local.messageBoxError, MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
-        private string GetStampImageLocation(string stamp)
+        private void LoadCurrentCarPreference(Cars car)
         {
-            switch (stamp)
+            Guna2Panel carPanel = new Guna2Panel()
             {
-                case "Audi":
-                    return "../../Images/CarsBrands/audi_logo.png";
-                case "BMW":
-                    return "../../Images/CarsBrands/bmw_logo.png";
-                case "Chevrolet":
-                    return "../../Images/CarsBrands/chevrolet_logo.png";
-                case "Ford":
-                    return "../../Images/CarsBrands/ford_logo.png";
-                case "Honda":
-                    return "../../Images/CarsBrands/honda_logo.png";
-                case "Hyundai":
-                    return "../../Images/CarsBrands/hyundai_logo.png";
-                case "Jeep":
-                    return "../../Images/CarsBrands/jeep_logo.png";
-                case "Kia":
-                    return "../../Images/CarsBrands/kia_logo.png";
-                case "Lexus":
-                    return "../../Images/CarsBrands/lexus_logo.png";
-                case "Mercedes-Benz":
-                    return "../../Images/CarsBrands/mercedes_logo.png";
-                case "Nissan":
-                    return "../../Images/CarsBrands/nissan_logo.png";
-                case "Subaru":
-                    return "../../Images/CarsBrands/subaru_logo.png";
-                case "Tesla":
-                    return "../../Images/CarsBrands/tesla_logo.png";
-                case "Toyota":
-                    return "../../Images/CarsBrands/toyota_logo.png";
-                case "Volkswagen":
-                    return "../../Images/CarsBrands/volkswagen_logo.png";
-                default:
-                    return "../../Resources/noAuto.png";
-            }
-        }
+                Size = new Size(265, 270),
+                Margin = new Padding(3, 3, 50, 75),
+                BorderColor = Color.Black,
+                BorderRadius = 30,
+                BorderThickness = 1
+            };
+            Guna2PictureBox carPictureBox = new Guna2PictureBox()
+            {
+                Size = new Size(265, 153),
+                FillColor = Color.White,
+                BorderRadius = 30,
+                BackColor = Color.Transparent,
+                CustomizableEdges = { TopLeft = true, TopRight = true, BottomRight = false, BottomLeft = false },
+                Location = new Point(1, 1),
+                SizeMode = PictureBoxSizeMode.StretchImage,
+                Image = Image.FromFile(car.Photo_1)
+            };
+            Guna2PictureBox carBrandPictureBox = new Guna2PictureBox()
+            {
+                Size = new Size(32, 32),
+                FillColor = Color.White,
+                BackColor = Color.Transparent,
+                Location = new Point(35, 172),
+                SizeMode = PictureBoxSizeMode.StretchImage,
+                Image = Image.FromFile(helper.GetStampImageLocation(car.Cars_Stamp.Stamp)),
+            };
+            Label carYearLabel = new Label()
+            {
+                Location = new Point(75, 192),
+                Font = new Font("Segoe UI", 10),
+                Text = car.Year.ToString(),
+            };
+            Label carNameLabel = new Label()
+            {
+                Location = new Point(75, 173),
+                Font = new Font("Candara", 12),
+                Text = car.Cars_Stamp.Stamp + ' ' + car.Cars_Model.Model,
+                Size = new Size(carPanel.Width - 85, 32),
+                TextAlign = ContentAlignment.TopLeft
+            };
+            Guna2CheckBox preferCheckBox = new Guna2CheckBox()
+            {
+                Size = new Size(90, 23),
+                Text = "Выбрано",
+                Font = new Font("Candara", 12),
+                Location = new Point(92, 231),
+                ForeColor = Color.FromArgb(160, 113, 255),
+                Checked = GetUserPreferences().Any(p => p.ID == car.ID),
+                Tag = car.ID
+            };
+            preferCheckBox.CheckedChanged += new EventHandler(preferCheckBox_CheckedChanged);
 
+            carPanel.Controls.Add(preferCheckBox);
+            carPanel.Controls.Add(carBrandPictureBox);
+            carPanel.Controls.Add(carYearLabel);
+            carPanel.Controls.Add(carNameLabel);
+            carPanel.Controls.Add(carPictureBox);
+            prefPanel.Controls.Add(carPanel);
+        }
         private void newListBtn_Click(object sender, EventArgs e)
         {
             LoadRandomCars();
@@ -217,10 +193,7 @@ namespace CarsAdviser.Forms
                     selectedPreferences.Remove(car);
                 }
 
-                if (!selectedPreferences.Contains(car))
-                {
-                    checkBox.Checked = false;
-                }
+                checkBox.Checked = selectedPreferences.Contains(car);
             }
         }
         private void SaveUserPreferences(Cars car)
@@ -229,57 +202,34 @@ namespace CarsAdviser.Forms
             {
                 using (var context = new AppContext())
                 {
-                    if (selectedPreferences.Count >= GetUserPreferences().Count)
+                    var existingPreference = context.Users_preferences.FirstOrDefault(up => up.Users_id == userId && up.Cars_id == car.ID);
+                    if (existingPreference == null && selectedPreferences.Contains(car))
                     {
-                        var existingPreference = context.Users_preferences.FirstOrDefault(up => up.Users_id == userId && up.Cars_id == car.ID);
-                        if (existingPreference == null)
+                        var userPreferences = new Users_preferences()
                         {
-                            var userPreferences = new Users_preferences()
-                            {
-                                Users_id = userId,
-                                Cars_id = car.ID
-                            };
+                            Users_id = userId,
+                            Cars_id = car.ID
+                        };
 
-                            context.Users_preferences.Add(userPreferences);
-                            context.SaveChanges();
-                            logger.Info($"Сохранение предпочтений пользователя для автомобиля с ID: {car.ID}");
-                        }
-                        else
-                        {
-                            logger.Warn($"Запись для автомобиля с ID: {car.ID} уже существует у пользователя с ID: {userId}");
-                        }
+                        context.Users_preferences.Add(userPreferences);
+                        context.SaveChanges();
+                        logger.Info($"Сохранение предпочтений пользователя для автомобиля с ID: {car.ID}");
                     }
-                    else if (selectedPreferences.Count < GetUserPreferences().Count)
+                    else if (existingPreference != null && !selectedPreferences.Contains(car))
                     {
-                        var existingPreference = context.Users_preferences.FirstOrDefault(up => up.Users_id == userId && up.Cars_id == car.ID);
-                        if (existingPreference == null)
-                        {
-                            var userPreferences = new Users_preferences()
-                            {
-                                Users_id = userId,
-                                Cars_id = car.ID
-                            };
-
-                            context.Users_preferences.Add(userPreferences);
-                            context.SaveChanges();
-                            logger.Info($"Сохранение предпочтений пользователя для автомобиля с ID: {car.ID}");
-                        }
-                        else
-                        {
-
-                            context.Users_preferences.Remove(existingPreference);
-                            context.SaveChanges();
-                            logger.Warn($"Запись для автомобиля с ID: {car.ID} уже существует у пользователя с ID: {userId}");
-                        }
+                        context.Users_preferences.Remove(existingPreference);
+                        context.SaveChanges();
+                        logger.Warn($"Удаление записи для автомобиля с ID: {car.ID}, так как он был удален из предпочтений пользователя с ID: {userId}");
                     }
                 }
             }
             catch (Exception ex)
             {
-                logger.Error($"Ошибка при сохранение предпочтений пользователя: {ex.Message}");
+                logger.Error($"Ошибка при сохранении предпочтений пользователя: {ex.Message}");
                 MessageBox.Show($"{Local.databaseConnectionError}: {ex.Message}", Local.messageBoxError, MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
+
         private void RemovePreferences()
         {
             try
@@ -484,7 +434,7 @@ namespace CarsAdviser.Forms
 
             similarToPreferences = AnalyzeUserPreferences();
 
-            parentForm.similarToPreferences = similarToPreferences;
+            UserPreferencesStorage.UpdateUserPreferences(similarToPreferences);
             MessageBox.Show(Local.preferencesTaken, Local.messageBoxInfo, MessageBoxButtons.OK, MessageBoxIcon.Information);
         }
         private void clearBtn_Click(object sender, EventArgs e)
